refactor(api): consolidate middleware directory structure under src/api/middlewares (#601)
