+++ src/api/routes/analyticsRoutes.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { track, bufferStatus } from "../controllers/analyticsController.js";
-import { authMiddleware } from "../../middleware/auth.js";
+import { authMiddleware } from "../middlewares/auth.js";
 
 const router = Router();
