+++ src/api/routes/bookmarkFolderRoutes.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { getFolders, createFolder, deleteFolder, organizeBookmark } from "../controllers/bookmarkFolderController.js";
-import { authMiddleware } from "../../middleware/auth.js";
+import { authMiddleware } from "../middlewares/auth.js";
 
 const router = Router();
