+++ src/api/routes/index.ts
@@ -19,7 +19,7 @@ import mentorshipRoutes from "./mentorshipRoutes.js";
 import bookmarkFolderRoutes from "./bookmarkFolderRoutes.js";
 import analyticsRoutes from "./analyticsRoutes.js";
 import recommendationRoutes from "./recommendationRoutes.js";
-import { errorHandler } from "../../middlewares/errorHandler.js";
+import { errorHandler } from "../middlewares/errorHandler.js";
 
 const rootRouter = Router();
 const v1Router = Router();
