+++ src/api/routes/karmaRoutes.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { getKarmaBalance, awardKarma } from "../controllers/karmaController.js";
-import { authMiddleware } from "../../middleware/auth.js";
+import { authMiddleware } from "../middlewares/auth.js";
 
 const router = Router();
