+++ src/api/routes/mentorshipRoutes.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { getMentorAvailability, bookSession, getSessions, updateSessionStatus } from "../controllers/mentorshipController.js";
-import { authMiddleware } from "../../middleware/auth.js";
+import { authMiddleware } from "../middlewares/auth.js";
 
 const router = Router();
