+++ src/api/routes/storageRoutes.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { handleSignatureRequest, handleSaveUpload, handleLocalUpload, localUpload } from "../controllers/storageController.js";
-import { authMiddleware } from "../../middleware/auth.js";
+import { authMiddleware } from "../middlewares/auth.js";
 
 const router = Router();
