+++ src/middlewares/validateRequest.ts
@@ -1,27 +1,4 @@
-import { Request, Response, NextFunction } from "express";
-import { ZodType, ZodError } from "zod";
-
-export const validateRequest = (schema: ZodType<any>) => {
-  return async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      const parsed = await schema.parseAsync({
-        body: req.body,
-        query: req.query,
-        params: req.params,
-      });
-      req.body = parsed.body;
-      req.query = parsed.query;
-      req.params = parsed.params;
-      return next();
-    } catch (error) {
-      if (error instanceof ZodError) {
-        return res.status(400).json({
-          success: false,
-          error: "Validation failed",
-          details: error.issues,
-        });
-      }
-      return next(error);
-    }
-  };
-};
+/**
+ * @deprecated Re-exported for backward compatibility. Import from `src/api/middlewares/validateRequest.js` instead.
+ */
+export * from "../api/middlewares/validateRequest.js";
